Create GitHub Actions for CI
This is intended as a replacement for AppVeyor and parts of TravisCI. It will create full builds of MAME (including tools) for the following platforms:

- Linux (GCC and Clang)
- macOS (Clang)
- Windows (GCC and MSVC 2019)

It will also upload the resulting executable as artifact so that it can be used for regression testing.

This commit also removes the unused TeaCI drone files.