Revet #7346.
The project won't move forward if we're not proactive about ensuring
legacy code isn't added.  In particular, this PR contained the kind of
code that we need to remove in order to be able to move to proper
per-device outputs and to allow clients to enumerate a system's outputs
on start; and runtime tagmap lookups for I/O ports, which can't be
checked on validation and hence require exercising all code paths to
detect errors (as well as the performance issues).

The pull request had outstanding reviews requesting changes.  It's not
acceptable to ignore that and merge it.

This reverts changes from e3dbb0fa43b78cf01e1f74b6dcc6fc8a7f4fc474
and fc8330e65cb0713a1931b7b3d771c37d5ffd020b.