k001005.cpp: Add missing command
Added a missing command only used in jetwave's "How to Control" screen.  This screen would show some model of the jetski in turning the handlebars and leaning on the jetski for controls. This fixes an issue where not only the model(s) wouldn't be present but the game would slow down complaining about an unknown polygon command.