qx10: Update keyboard to match HASCI layout
This updates the keyboard for the QX10 to better match a single layout.
Prior to this it tried to be both the US ASCII layout and HASCI layout
at the same time. This causes some issues since some keys such as
Caps Lock/Shift Lock and Delete are on different rows/columns between layouts.

In addition this also maps in a few previously missing keys and adjusts some
other key mappings,