stfight.cpp: Remove YM2203 frequency hack, and set prescaler at machine_start
Empire City: 1931 has implemented a hack to 3x YM2203 clock frequency as a solution to pitch of all sounds being too low. But FM (BGM) be correct, but SSG (timer countdown, clear bonus count) be too high pitch.
Frequency hack can't seem to get the correct pitch.
This PR removes frequency hack and sets YM2203 clock prescaler at machine_start in order to get right pitch.
This PR may be a hack-like approach, but I proposed the fix in this because this can remove hack that result in wrong pitch.