Add (p)atan to MAME softfloat
These are sourced from Bochs like the fsincos one, adapted to work
with MAME (change symbol prefix, change includes, remove status,
C99 stdint types, macro massaging, copying from other files in the
subdirectory like fsincos/specialize/fyl2x, symbol changes).

Because this is FPATAN, we should be able to simulate ATAN by
hardcoding one as one of the arguments.

This could be useful for MAME's x87 emulation as well.

Since the licensing is the same, the file should be no problem.
However, a lot of the code is duplicated and probably could be
merged into a common header; I'm unsure of the best strategy for that.