cyclemb.cpp: Remove MACHINE_IMPERFECT_SOUND flags
I think Cycle Maaboh sounds was corrected on pull request for #7979.
And, Sky Destroyer does not seem to have any sound imperfections compared to PCB.
So I want to remove the MACHINE_IMPERFECT_SOUND flag from the driver.
(If you think I should not remove the flag, please review it)