pofo: Create real keyboard device
The current implementation had a severe flaw - it did not properly
support multiple keys being pressed down at the same time, but only
registered the key that was the "first" in the scan matrix.

This lead to being unable to type certain capital letters, some
shortcuts not working etc.

Fix this by creating a real keyboard device based on the common
`device_matrix_keyboard_interface` class.