dai3wksi.cpp: Correct Runaway title.
The advertisements prominently show “RUNAWAY” in uppercase Latin script
with no space.  Additionally, the title is rendered as “ランナウェイ” in
Katakana.  Splitting into on, this is “ra-n-na-we-i” - note the “na”,
indicating there’s no break between the consonant and vowel.  If it was
supposed to be two words, it would more likely be rendered
“ランアウェイ” (“ra-n-a-we-i”).