acrnsys.cpp: change keyboard default to the real keyboard (MT#08020)
The root system has a 1-key “keyboard” for the break key so MAME disables the second keyboard (the real keyboard) by default.  This changes makes the real keyboard the default keyboard which removes the need to change to the real keyboard within the UI.  This change is an exact copy of the one line change proposal by Vas Crabb seen on MT#08020.