MAKE_INT_8 fails on GCC/ARM
The old MAKE_INT_8 macro failed (added 0x100 displacement in the BRANCH method) on gcc-arm-none-eabi-10-2020-q4-major since by default __CHAR_UNSIGNED__ is defined on that platform. Now there might be a better solution to this. Figured this is 2021 and int8_t should be available everywhere. Fixes works on MSVC2017, (gcc-arm) and XCode12 (llvm). Now MAME may use -fsigned-char when compiling and this change makes no sense. I am using this file outside and took me a while to find this issue.