lwriter.cpp: Reimplement the 0x600000-0x7fffff range to bitwise-OR on write.
This appears to be the desired effect of the code using this range.
It fixes the rasterization of the '5' glyph in the test page and avoids
the "Fatal system error @ 0x2a8416" buffer overflow in glyph the cache
when writting out the last byte of the bitmap because it ends up or'ing
with zero instead of overwritting with zero.

Fixes #8309