lwriter.cpp: Invert the state of the SCC interrupt
This dramatically improves serial output speed, which I believe
was limping along off the timer interrupt. It also fixes serial output
on channel B as well as with the ROM for the original LaserWriter.

Further, it better matches the 0x60 that VIA_PCR is set to which has
CA1 latch/interrupt control set to "Negative active edge".