New drivers: Technics KN5000 and KN1500 musical keyboards
----
New TLCS900 variant: TMP94C241

This also differentiates the disassemblers so that they can support the different sets of special function registers (SFRs) of each cpu variant.

----
new non-working devices: kn1500 & kn5000 Technics keyboards

----
many many many SFRs + timers + AD + interrupts...

mostly based on code duplication from tmp95c063
but with many adaptations to fit the descriptions of the tmp94c241 datasheet.

in the future the duplicated code could possibly be moved to a parent class of both devices, but for now I accepted to make this duplication so that I could at least wrap my head around this and start to get comfortable with the CPUs differences as well as similarities. This knowledge will help me to refactor it later if needed.

----
Improved the implementation of 8bit timers

----
fixed TLCS900 / TMP94C241 disasm of control-regs but break the disasm of the other CPUs in the family. So I'll need to make an additional commit to actually support these disasm differences accross the family.