Correctly count roms not found.
The counting of roms not found doesn't output correctly.  These changes will count the not found roms correctly and add a not found count to the output.

To see what I mean,  Take original Mame, remove (or rename temporarily) a romset and do a verifyroms on that set.  Then do it with my changes implemented.