util/delegate.h: Use "compatible" delegates for MinGW GCC i686.
The Itanium delegate has questionable value on on MinGW i686 as it
injects a conditional branch in the hot path for delegates anyway to
deal with the different __thiscall convention.  It's somehow breaking
and causing memory corruption in full builds, but I don't have the time
to work out exactly which delegate type is the problematic one,
especially not with a release coming soon.

This will probably cause 32-bit MinGW builds to become substantially
larger.