Implemented address spaces syntax using colong delimiters only.
Switched pcatmem family of commands to accept device/space as an
optional qualifier on the address argument.

Fixed the cpunum symbol so it uses the same rules for indexing devices
as cpulist and the argument parser.

Made device_t sanitise subtags better.  Previously you could cause an
assertion failure or crash MAME by giving it unexpected relative tags
from Lua or the debugger.

Did a quick sweep of the command syntax summaries in the debugger help.
Descriptions are not updated yet.