frontend: Recovered precious vertical space for system/software names!
Reduced the height of the bottom info panel by one line.  The system
shortname or list/software name tuple are now only displayed in the info
box on the right.  It's a bit out-of-the-way, but it's not something you
need to see all the time.

The main reason for having the quit or return to previous menu item
always visible is to make it easy for someone to quit if they only have
a mouse.  However, we don't need waste space on the menu for this when
we have a toolbar.  Users without a mouse can exit using the
keyboard/controller UI Cancel input (given how important this is, it's
unlikely they won't have it mapped to something usable).

There's now a toolbar button on the extreme right for returning to the
previous menu or quitting.  The tooltip and icon are appropriately
context-sensitive.  This recovers one line on the system selection menu,
and two on the software selection menu since the separator is no longer
needed.

Replaced the toolbar icons bitmaps with SVGs.  Colours/shapes may be
tweaked before the release if people have good suggestions

Flipped the simple system selection menu so the description is on the
left.

Moved the handling of special cases for the final menu item out of the
base menu class.  It's still hacky having it handled there at all, but
it's less hacky without if menus that want to do something different can
do it themselves.