frontend: Cleaned up rendering of info views.
Put the description for systems in the info box - it's useful for the
fruit machines with very long names that are truncated in the list.
Also stopped truncating manufactuer and parent name in the info box.

Made the text layout class capable of handling lines containing
combinatations of left/centre/right-justified text and got rid of the
legacy UI manager text wrapping function.  Made the system/software
selection menus and the info viewer share the same code for formatting
info text.  This means the multi-column layout works properly in the
info viewer now, and the code is a lot simpler.  Also the
system/software selection menus don't have to redo the text layout every
frame now.

Made the info viewer update the text layout if the output aspect ratio
changes, and cleaned up more legacy code.  The lines in the info viewer
are no longer bogus "menu items", and there's a lot less special-case
code to support it in the base menu class.

This commit includes an update to the Chinese translations from YuiFAN.