shootgal: Adjusted MAME's crosshair to match actual crosshair (#8739)
Shooting Gallery (shootgal)'s default PORT_CROSSHAIR values are reverse and unscaled against the game's screen resolution.

The game does not offer a calibration option, but the test mode that comes up after a few seconds upon setting the DIP switch shows the internal crosshair for adjusting the gun.

The new offsets are manually determined. They are visual cues for the MAME users only and are unrelated to the game's operation.