Build fix on mac: 10 ^ 7 hits the -Wxor-used-as-pow check
On building with recent clang on mac, 10 ^ N hits -Wxor-used-as-pow
check. This happens only when a left value is 10, and we can explicitly
say xor to express an exclusive OR operation.

This simple patch replace existing '^'s with 'xor's so that we can
compile it on recent mac, and the code looks consistent in the file.