ksys573.cpp - Fixed save state bug preventing stage input
Changed the init_ddr to remove the GX700PWFBF call and place that in init_ddra so that analogue mixes still used it. This fixes a bug that when using save states, the stage io would not be saved and so would be unplayable.

Also changed input labels to be consistent with other versions of DDR.