daiohp2: incorrect bonus life DSW options
daiohp2 bonus life DSW options in this set are similar to the final release, and different than the other prototype set (daiohp). Now fixed.

That makes me think that daiohp2 is in fact a recent set closer to the release set than the other prototype set, a not an earlier one as now stated. But that needs to be analyzed in more depth.