galaxian.cpp: Identified three inputs for thepitm (MT08280)
Unusually, this game considers a single coin to be a single credit which means you could need 2 credits (coins) before a game can start depending on the Coinage DIP switch setting.  I have tried to come up with suitable descriptions for the Coinage DIP switch to avoid confusion.  Note that the Coinage DIP switch on the original thepit set is also affected which I will capture in a separate PR.  If there is a better description to use, please let me know.  Thanks in advance.