Consolidated floppy_image_device::m_create_fs and floppy_image_device::m_io_fs vectors
We had two separate memebers in floppy_image_device (m_create_fs and m_io_fs) that contained the same data.  Whether the file systems can be formatted or read can be identified by querying fs::manager_t.

For this reason, it seems bad to have these separate members, the seemingly only reason to make the UI code slightly less complicated.  For this reason I consolidated these two members and moved the burden of selecting which ones are creatable to the UI code.