Fixed some issues involving fs::meta_description constructing
With the recent change to use std::variant more closely, I noticed a problem where meta_descriptions of type meta_type::String got defaults of type 't'.  This was because the templated ctor for meta_description would convert 'const char *' to 'bool'.  This change adds another overload to catch 'const char *', along with asserts to catch problems.

In the process I corrected a few meta_description ctors

It is possible that this change does not go far enough.  Perhaps the meta_type argument should be removed, and we should instead create distinct ctor types (rather than relying on templates) and specify the precise meta_type in the overload.  Or even go further and remove m_type from meta_description, and instead create an overload that calculates meta_type based on calling std::visit on the variant