Added support for specifying volume attributes in 'floptool flopcreate'
An example command line:

  flopcreate vdk coco_rawdsk_os9_35 newdisk.vdk -name mycooldisk -creation_date "1999-02-28 13:23:47"

Attributes are identified on the command line prefixed with '-'; if this is not the preferred syntax this can be changed.

Implementing this also forced a change to fs::meta_value where the various as_*() calls can now be called without respect to which type the fs::meta_value is; this is necessary so that floptool code doesn't need to "own" parsing of the various types of fs::meta_value.

And with this change, fs::meta_value::to_string() is now replaced by fs::meta_value::as_string()