Reporting error when software list names are ambiguous
MAME supports running software from the command line without specifying the list.  This can be a problem when there are multiple software list items that are ambiguous.  For example:

	`mame expert11 hero`

Could be either:

	`mame expert11 msx1_cart:hero`

or:

	`mame expert11 msx1_cass:hero`

This changes MAME's logic so that any ambiguity results an error being raised at the command line.  Items identified as "required" that are not fully qualified are now similarly expected to be unambiguous.

There are also other oddities about the current logic that I did not attempt to change.  For example, the logic seemed to only require that a single part be compatible to identify all software as compatible.  This seems incorrect but I'm not brave enough to change this.

Lastly, this was a non-trivial refactoring of old code, so here are test cases that were identified:
- `mame expert11 hero`; fails with an error that the request was ambiguous
- `mame expert11 msx1_cart:hero`; succeeds
- `mame expert11 msx1_cass:hero`; succeeds
- `mame expert11 msx1_cass:zyxzyx`; fails
- `mame expert11 zyxzyx`; fails
- `mame c64 c64_flop_misc:mvdemo` (has fully qualified required software in a different list); succeeds
- `mame pce aiii` (has partially qualified required software in a different list); succeeds