8080bw.cpp: Improved emulation for Sidam Invasion. [Janniz]
One of the R Z SRL Bologna sets was aleady working as well as any of the
other sets.  This improves inputs for all of them anyway, so may as well
credit Janniz for it even though it's not clear when it started working.

Clones promoted to working
--------------------------
Invasion (bootleg set 1, R Z SRL Bologna) [Janniz]