Fixing fs::cbmdos_image::enumerate_f() to properly look for single density (instead of double) (#10104)
This problem was identified because the file system didn't properly show up when the 'c64' driver was run, and this seems to be the correct fix.  The part of this that confuses me is that if this defect was preventing the CBMDOS fs from showing up under c64, why don't the CoCo file systems have the same problem?  Both fs_coco_rsdos.cpp and fs_coco_os9.cpp specify (seemingly incorrectly) SSDD.