Changed fs::manager_t::enumerate_f() to simplify logic within file system drivers
enumerate_f() used to contain quite a bit of boilerplate logic to determine whether a particular floppy type should be added.  This change attempts to move this logic outside the file system drivers to simplify the drivers.

The riskiest part of this change is unformatted_image::enumerate_f().  I attempted to replicate the logic that was previously determining with unformatted image types to use, but the logic is tricky and it isn't clear to me that replicating the logic is the correct action - I may be cargo culting.