mastboy.cpp: Dumped and added a newer Italian set
Also added comments and reformatted existing single line ones.
Added placeholders for PLDs.
The Italian sets are not hacks. Gaelco licensed the game to Playmark for the Italian market, so changed the manufacturer accordingly on the Italian sets.
Finally, the Rev A on the piggyback PCB refers to the game revision, not the PCB revision, so, changed the games names removing the "PCB" reference.