Adding CP/68 Disk format for SWTPC 6800
CP/68 is an operating system for the SWTPC 6800 computer which was
published in a book as well as available commercially from the author.
Because the operating system source code was published in the book the
code was OCRed thus reviving the operating system.  This code change
adds the CP/68 disk format which was adjusted by Mike Douglas of
deramp.com to work on real hardware.  The same disk images he published
work with this code change unaltered.