bus: Get rid of some dubious tag manipulation.
The implementation details of how the cartridges allocate storage for
memory really shouldn't be part of the interface.  Having tags in the
headers encourages people to depend on these implementation details.
This gets rid of it in most of the headers.  A few particularly leaky
abstractions (bbc/rom, electron/cart, gba, generic, jakks_gamekey, m5)
depend on this, so it can't be removed in those cases without further
refactoring to encapsulate the slot devices better.

This doesn't change behaviour, it just mechanically removes stuff from
the headers and uses device_t::subtag rather than string manipulation on
tags.  Most of the cartridge devices shouldn't have rom_alloc member
functions at all - the region created by the software list loader can be
used directly when loading from the software list, and the slot can
allocate a region with the same tag when loading loose software.  This
avoids creating an extra region and copying the data when loading from
the software list.  See vboy for an example that doesn't allocate a
superfluous region.