Revert "gen_latch: add boost interleave conf setting"
Jamming magic behaviour for a few systems into a common device just
makes code more convoluted and harder to maintain.  The trouble with
these things is that you quickly end up stuck with them, and then
untangling the dependency on magical behaviour that never belonged in
the device in the first place becomes an inordinate amount of work.

In this case, it isn't even saving a substantial amount of code in the
drivers.

This reverts commit 229ea4b3a1b44bd558e2c99c31dfe3ad66c34ea0.