Mark new bosco parent set PROMs as good
I was going to dump this version a few months ago because I got a PCB with it, but I forgot about it until now. Anyways, I can verify that the PROMs are the same as on the other revisions. The only thing of note is that BOS1-3 and BOS1-4 on my PCB have swapped labels, but the data was correct for the board location. Considering that my BOS1-6 also has an upside label, I am going to assume that it was just a label mistake.