newport: Split out iterator setup stage (#10546)
In REX3, iterator setup should be a dedicated operation that
is only execuated if
a). A Draw command is execuated with DoSetup at DRAWMODE0 set
or
b). A host write is issued to SETUP register

At setup stage REX3 will calcuate quadrant for a block or span
drawcall or octant together with some Bresenham parameters for
a line drawcall.

Linux newport_con driver is rely on this behavior to use quadrant
calculated by previous draw call to render characters.

Fixes: #9667
Signed-off-by: Jiaxun Yang <jiaxun.yang@flygoat.com>

Signed-off-by: Jiaxun Yang <jiaxun.yang@flygoat.com>