skeleton: Rename Vector 4 to Vector 3
It appears Vector Graphic Monitor 4 was confused with Vector 4. The
majority of the functioning parts are for the Vector 3, as that is the
source of the ROMs.

I've fixed some of the easier Vector 4-isms that were copied from the
user guide, but it will need more thorough work. That should be a lot
easier with the better understanding of which machine was running and
the detailed documentation.