Revert "Make numbers signed in debugger where it matters"
This reverts commit 922d5b3c81459173ce99970dca0e44825f0a4494.

This broke an important use case with MIPS3 address comparisons no
longer behaving intutively.

It's also another case of a debugger change that wasn't followed through
properly, with absolutely no attempt to update the internal debugger
help or the documentation for the web site.

I realise signed maths woud be a useful feature, for the debugger, but
it needs to be done in a way that doesn't break important use cases.