nintendo/punchout.cpp: Swap left and right speakers.
Schematics have 2A03 audio going to MSOUND and VLM5030 going to SSOUND on the P5 edge connector. These in turn are routed to the sound boards on the lower and upper monitors respectively. Provided left speaker is connected to the lower monitor's boards and vice versa this change should be correct. The left/right layout of the diagram in the operator's manual also implies this positioning.