sega/sms.cpp: Refactored Game Gear handling.
Use a memory view to switching between Game Gear and Master System
compatibility I/O space mapping.  Got rid of most of the conditional
code checking whether it's a Game Gear driver.

Got rid of the special Game Gear EXT port slot.  It's the same thing as
a Mega Driver controller/expansion I/O port with a different connector.

sega/mdioport.cpp: Added a variant for the Game Gear with the TH/PC6
interrupt latch, and added it to the Game Gear.