Apple II double lo-res: work around issue #10760
This implements the same workaround for double lo-res that was already
used for double hi-res: shift the whole image right by one pixel (losing
the rightmost pixel) in monochrome mode so that the NTSC shader is
happy. This doesn't affect color mode.

This rewrite incidentally fixes an incorrect shift count in the old code
(see the issue).