Improved Apple II hires graphics color simulation (#10773)
This algorithm decodes the scan line to 560 bits in the same way as the
hardware (and MAME's monochrome modes) and uses a 7-bit sliding window
on the decoded bits. The result is the same as before if all high bits
are clear. If high bits are set, the output is shifted right by one
pixel as it should be, and the unique artifact colors at the boundaries
between bytes with opposite high bits are simulated. (For more about
those colors see page 8-20 of Understanding the Apple II by Jim Sather.)

The monochrome modes are rewritten to use the same decoding logic, which
incidentally fixes some bugs: the left pixel of a 14-pixel group was not
updated at 0-to-1 bit-7 transitions (leaving old pixels on the screen),
the test against cliprect was incorrect, and m_dhires was ignored.