Better Apple II artifact color table (#10792)
Replace the "old" (committed just a few days ago) artifact color lookup
table, which was designed in a somewhat ad-hoc way for hires graphics
only, with a table based on simpler rules that can be used in any Apple
II (not GS) video mode, though at the moment it is still used only for
hires. Move it upward in the file in anticipation of other uses.