Apple II: fix Video-7 RGB card double hi-res modes
Fix an off-by-1 error in the 160-column mode that caused the rightmost 2
pixels to be omitted. In 140-column color and mixed modes, use custom
logic to make 140 wide pixels instead of simulating blurry artifact
color. In mixed mode, each high bit controls 7 narrow pixels according
to the manual.

With these changes, the card seems to pass all of the tests on the
official demo disk, except that foreground-background hires remains
unimplemented, "Color Mouse Draw" seems to hang/crash, and some demos
sometimes set the wrong mode when run after other demos (including
themselves). The last could be a bug in the demos.