Separate Apple II video config from system config
Add a a2_video_config port to contain the monitor type (with the idea of
adding other settings later), and add some private accessors to avoid
magic numbers. This incidentally fixes a bug in apple2 and apple2cp
which had another configuration field that overlapped with the
monitor-type field. On the IIc+, setting 80-column mode would make the
video system think a IIe RGB card was installed.