Changed CoCo Raw format to be anonymous
This is an experimental change to make the "CoCo Raw Disk" format be anonymous; not to be exposed to external users but rather to be an intermediate format for use by the CoCo file system modules.  In practice, this was quite easy - it was just a question of removing FLOPPY_COCO_RAWDSK_FORMAT from global listings and changing the various methods that only need to exist for public formats (name/description/extensions) to return nullptr.

I do not think that this is the ideal approach.  For anonymous formats name/description/extensions become farcical baggage and a better approach might be to create a base class (e.g. - anonymous_floppy_image_format_t) that omits this baggage.  I'm looking forward to hearing ideas!