rgbgen: Safely compute out of range shifts
The SSE and VMX implementations of rgbaint_t have consistent and well
defined behavior for out of range shifts: the shift amount is treated as
unsigned and shifts equal to or greater than the element width either
zero the result (shl, shr) or replicate the sign into every bit (sra).

Shifts of this amount are undefined behavior in C++ and so extra checks
have been added to reproduce the same behavior in the generic
implementation. The validity checker has also been updated to add cases
for out of range shifts.

Some explicit sign extensions have been removed from sra() as the
validity checker already requires that signed right shifts must be
arithmetic.

This bug was encountered when building the N64 RDP for ARM64 as it
caused very obvious rendering artifacts.